Create new components to make the code more cleaner
